+++ jtool/utils/debug.py
@@ -1,0 +1,12 @@
+DEBUG_ENABLED = False
+
+
+def enable_debug():
+    global DEBUG_ENABLED
+    DEBUG_ENABLED = True
+
+
+def print_debug(*args):
+    global DEBUG_ENABLED
+    if DEBUG_ENABLED:
+        print(*args)
