+++ jtool/utils/func_asserts.py
@@ -1,18 +1,17 @@
-from .utils import raise_error, json2shortstr
+from ..utils.errorhandling import raise_error
 
 
 def lambda_type(data, ttype):
     if isinstance(data, ttype):
         return data
-    raise_error(json2shortstr(data), "selection is not of " + str(ttype))
+    raise_error(data, "selection is not of " + str(ttype))
 
 
 def lambda_multi_type(data, *test_types):
     for t in test_types:
         if isinstance(data, t):
             return data
-    raise_error(json2shortstr(data),
-                "selection is not of " + str(t))
+    raise_error(data, "selection is not of " + str(t))
 
 
 def lambda_member(item, collection):
@@ -24,4 +23,4 @@ def lambda_member(item, collection):
     if isinstance(collection, dict):
         collection_str = "dict"
     e_message = str(item) + " not in " + collection_str
-    raise_error(json2shortstr(collection), e_message)
+    raise_error(collection, e_message)
